% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaODE.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach: Ensure Julia Dependencies on Package Load}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library name (automatically provided by R when the package is loaded).}

\item{pkgname}{The package name (automatically provided by R when the package is loaded).}
}
\description{
This function is automatically called when the package is attached to the R session.
It ensures that all required Julia dependencies are installed and ready for use.
}
\details{
The \code{.onAttach} function ensures that all required Julia dependencies are installed
by calling \code{ensureJuliaPackages}. It provides startup messages to inform the user
about the status of the installation process.
}
\note{
Users do not need to call this function directly; it is executed automatically when
the package is loaded using \code{library()} or \code{require()}.
}
\seealso{
\code{\link{ensureJuliaPackages}}
}
