% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaODE.R
\name{ensureJuliaPackages}
\alias{ensureJuliaPackages}
\title{Ensure Required Julia Packages Are Installed}
\usage{
ensureJuliaPackages()
}
\description{
This function checks for and installs any required Julia packages needed by the R package.
}
\details{
The function sets up the Julia environment using the \code{JuliaCall} package, writes a temporary
Julia script with functions to manage package installation, and ensures that required Julia
packages are available. If any package is missing, it will be installed automatically.
}
\note{
This function is called internally by \code{.onAttach} and does not need to be called directly
by the user.
}
\seealso{
\code{\link{.onAttach}}
}
