% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaODE.R
\name{ensureJuliaPackages}
\alias{ensureJuliaPackages}
\title{.onAttach: Ensure Julia Dependencies on Package Load}
\usage{
ensureJuliaPackages()
}
\arguments{
\item{libname}{The library path to the package.}

\item{pkgname}{The name of the package being attached.}
}
\description{
This function is called automatically when the package is attached to the R session
(e.g., using \code{library(yourpackage)}).
}
\details{
The \code{.onAttach} function ensures that all required Julia dependencies are installed
by calling \code{ensureJuliaPackages}. It provides startup messages to inform the user
about the status of the installation.
}
\note{
Users do not need to call this function directly; it is executed automatically.
}
\seealso{
\code{\link{ensureJuliaPackages}}
}
